Delete duplicated class
