--- src/test/java/PawnTest.java
@@ -1,24 +1,0 @@
-import chess.pieces.Pawn;
-import org.junit.jupiter.api.Test;
-import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-public class PawnTest {
-
-    @Test
-    public void create() {
-        verifyPawn(Pawn.WHITE);
-        verifyPawn(Pawn.BLACK);
-    }
-
-    @Test
-    public void create_기본생성자() throws Exception {
-        Pawn pawn = new Pawn();
-        assertEquals(Pawn.WHITE, pawn.getColor());
-    }
-
-    void verifyPawn(final String color) {
-        Pawn pawn = new Pawn(color);
-        assertThat(pawn.getColor()).isEqualTo(color);
-    }
-}
